Backend/requirements.txt
flask
flask-cors
tensorflow
pillow
rembg
numpy
